Research pivot: config comparison, E2E latency, YOLO CPU baseline
Advisor feedback (March 23): shift from isolated benchmarks to
system-level configuration comparison with E2E latency as headline metric.

YOLO + ViSP (Ahmed):
- visp_servo_node.cpp: add detection-header-stamp → cmd_vel E2E latency
  measurement; logs 'e2e=XXms' alongside existing callback latency
- yolo_cpu_producer.py: new CPU inference producer (Ultralytics PyTorch);
  drop-in replacement for Hailo NPU path, same SHM interface
- benchmarks/e2e_latency_probe.py: ROS2 probe node for det→cmd_vel latency
- benchmarks/contention_bench.sh: perf stat profiling across 4 conditions
- benchmarks/yolo_cpu_bench.sh: benchmark all YOLO26 sizes on CPU

SLAM (Amar):
- ov2slam_ros: precision_config.hpp, frame.hpp/cpp float32 bearing vectors
- FE speedup configs: fe_klt2, fe_maxdist65, fe_combined yamls
- aggregate_slam_results.py, plot_paper_figures.py, plot_ipc_cdf.py,
  core_pinning_bench.sh, ipc_latency_bench.py, plot_resolution.py,
  yolo_fp_breakdown.sh, yolo_nms_comparison.sh, yolo_resolution_bench.sh

Cleanup:
- Remove gstshark calibration trace (camera_calib/)
- Add benchmarks/results/ and .claude/ to .gitignore

+++ src/ov2slam_ros/src/loop_closer.cpp
@@ -473,10 +473,10 @@ bool LoopCloser::epipolarFiltering(const Frame &newkf, const Frame &lckf, std::v
 
     for( const auto &kplmid : vkplmids ) {
         auto kp = newkf.getKeypointById(kplmid.first);
-        vcurbvs.push_back(kp.bv_);
+        vcurbvs.push_back(kp.bv_.template cast<double>());
 
         auto lckp = lckf.getKeypointById(kplmid.second);
-        vlcbvs.push_back(lckp.bv_);
+        vlcbvs.push_back(lckp.bv_.template cast<double>());
     }
 
     bool success = 
@@ -796,7 +796,7 @@ bool LoopCloser::p3pRansac(const Frame &newkf, std::vector<std::pair<int,int>> &
         auto kp = newkf.getKeypointById(kpid);
 
         vwpts.push_back(plm->getPoint());
-        vbvs.push_back(kp.bv_);
+        vbvs.push_back(kp.bv_.template cast<double>());
     }
 
     int k = 0;
